+++ src/core/renderer_mtl/renderer_mtl.cpp
@@ -404,7 +404,6 @@ void RendererMTL::drawVertices(PICA::PrimType primType, std::span<const PICA::Ve
 	renderCommandEncoder->setVertexBytes(&regs[0x48], 0x200 - 0x48, 0);
 	renderCommandEncoder->setFragmentBytes(&regs[0x48], 0x200 - 0x48, 0);
 
-	// TODO: respect primitive type
 	renderCommandEncoder->drawPrimitives(toMTLPrimitiveType(primType), NS::UInteger(0), NS::UInteger(vertices.size()));
 }
